chore: Update dependencies and global styles
Refactor the home page component and update the global styles to use the new version of Next.js (14.2.3) and add the "react-icons" dependency (v5.2.1). Also, import and use the "Navbar" component in the "_app.js" file for consistent navigation across all pages. Additionally, update the font family and styles in the "globals.css" file to use the "Baloo Bhai 2" font from Google Fonts. Finally, configure the Tailwind CSS container to be centered with padding.